+++ app/src/main/java/com/ysshin/fine_dust_app/data/RegistrationRepository.kt
@@ -1,8 +1,6 @@
-package com.ysshin.fine_dust_app.repository
+package com.ysshin.fine_dust_app.data
 
 import com.ysshin.fine_dust_app.api.AuthService
-import com.ysshin.fine_dust_app.data.AuthData
-import com.ysshin.fine_dust_app.data.RegistrationData
 import retrofit2.Call
 
 class RegistrationRepository(private val authService: AuthService) {
