+++ app/src/main/java/com/ysshin/fine_dust_app/viewmodel/RegistrationViewModel.kt
@@ -1,18 +1,12 @@
 package com.ysshin.fine_dust_app.viewmodel
 
-import android.util.Log
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import androidx.navigation.fragment.findNavController
-import com.ysshin.fine_dust_app.R
 import com.ysshin.fine_dust_app.api.AuthService
 import com.ysshin.fine_dust_app.data.AuthData
-import com.ysshin.fine_dust_app.data.PreferenceManager
 import com.ysshin.fine_dust_app.data.RegistrationData
-import com.ysshin.fine_dust_app.repository.RegistrationRepository
+import com.ysshin.fine_dust_app.data.RegistrationRepository
 import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
 class RegistrationViewModel(authService: AuthService) : ViewModel() {
